Add files via upload
I have done standard scaling, outlier removal, label encoding of categorical columns. Linear regression is the model used to calculate the accuracy. And it has been concluded that linear regression model is having very low accuracy to determine the model performance and hence it adversely affects the price prediction. Moreover there are no major factors which can be considered as price prediction factors as none of the factors are  showing high correlation with the target variable. Some other features need to be added to the the dataset to improve accuracy.